+++ api/src/test/java/org/grnet/status/api/UserEndpointTest.java
@@ -3,10 +3,9 @@ package org.grnet.status.api;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.http.ContentType;
-import jakarta.inject.Inject;
 import org.grnet.status.api.endpoints.UserEndpoint;
+import org.grnet.status.dtos.InformativeResponse;
 import org.grnet.status.dtos.user.UserProfileDto;
-import org.grnet.status.services.UserService;
 import org.junit.jupiter.api.Test;
 
 import static io.restassured.RestAssured.given;
@@ -15,10 +14,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @QuarkusTest
 @TestHTTPEndpoint(UserEndpoint.class)
 public class UserEndpointTest extends KeycloakTest {
-
-
-    @Inject
-    UserService userService;
 
     @Test
     public void getUserProfile() {
@@ -39,5 +34,20 @@ public class UserEndpointTest extends KeycloakTest {
 
         assertEquals("LOCALTENANT", group.name);
         assertEquals("viewer", group.role);
+    }
+
+    @Test
+    public void registerUser() {
+        var response = given()
+                .auth().oauth2(tenantViewer)
+                .contentType(ContentType.JSON)
+                .post("/registration")
+                .then()
+                .assertThat()
+                .statusCode(200)
+                .extract()
+                .as(InformativeResponse.class);
+
+        assertEquals(response.message, "Registration completed.");
     }
 }