+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/GroupMembersResponse.java
@@ -1,0 +1,8 @@
+package org.grnet.status.authorizations.dtos;
+
+import java.util.List;
+
+public class GroupMembersResponse {
+
+    public List<GroupMemberEntry> results;
+}