+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/GroupUser.java
@@ -1,0 +1,10 @@
+package org.grnet.status.authorizations.dtos;
+
+public class GroupUser {
+
+    public String id;
+    public String username;
+    public String firstName;
+    public String lastName;
+    public String email;
+}