+++ authorization/src/main/java/org/grnet/status/authorizations/tokens/AccessTokenProvider.java
@@ -1,0 +1,5 @@
+package org.grnet.status.authorizations.tokens;
+
+public interface AccessTokenProvider {
+    String getAccessToken();
+}